Refactor bidcycle views, add bid object
