+++ requirements.txt
@@ -7,6 +7,7 @@ coverage==4.4.1
 defusedxml==0.5.0
 dj-database-url==0.4.2
 Django==1.11.3
+django-choices==1.6.0
 django-cors-headers==2.1.0
 django-debug-toolbar==1.8
 django-extensions==1.7.9
@@ -28,6 +29,7 @@ openapi-codec==1.3.1
 pbr==3.1.1
 psycopg2==2.7.1
 py==1.4.34
+pyflakes==1.6.0
 pytest==3.1.2
 pytest-cov==2.5.1
 pytest-django==3.1.2
