+++ talentmap_api/bidding/views/bidcycle.py
